requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
--find-links https://data.pyg.org/whl/torch-2.5.1+cu121.html
torch==2.5.1+cu121
torch-geometric==2.6.1
torch_cluster==1.6.3+pt25cu121
torch_scatter==2.1.2+pt25cu121
torch_sparse==0.6.18+pt25cu121
torch_spline_conv==1.2.2+pt25cu121
torchaudio==2.5.1+cu121
torchmetrics==1.6.1
torchvision==0.20.1+cu121
h5py==3.12.1
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
mhfp==1.9.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
munkres==1.1.4
numba==0.60.0
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
pyarrow==18.1.0
pytorch-lightning==2.5.0.post0
PyYAML==6.0.2
rdkit==2024.9.4
scikit-learn==1.5.2
seaborn==0.13.2
sympy==1.13.1
tornado==6.4.2
tqdm==4.67.1
wandb
pandas